add check user id in cart clear

+++ Product service/controllers/cartController.js
@@ -290,6 +290,12 @@ const removeCart = async (req, res, next) => {
 const clearCart = async (req, res, next) => {
     let userId = req.params.userId
     let cart = await Cart.findOne({user_id: userId})
+    if(!cart){
+        return res.status(400).json({
+            success: false,
+            message: "user ID is not exist"
+        })
+    }
     cart.total_item = 0
     cart.total_price = 0
     await CartDetail.deleteMany({cart_id: cart.cart_id})
